# Conflicts:
#	src/main/java/com/team/student_calendar/controller/api/BookApiController.java
#	src/main/java/com/team/student_calendar/entity/BookEntity.java
#	src/main/java/com/team/student_calendar/service/book/SelectBookService.java

+++ src/main/java/com/team/student_calendar/repository/BookRepository.java
@@ -9,5 +9,5 @@ import java.util.List;
 @Repository
 public interface BookRepository extends JpaRepository<BookEntity, Long> {
 
-    List<BookEntity> findAllByBookNoIn(List<Integer> bookNoList);
+    List<BookEntity> findAllByBookNoIn(List<Long> bookNoList);
 }
